Refactor `generateDefaultNotebookName` into utils
