+++ src/index.ts
@@ -19,7 +19,7 @@ import { Commands } from './commands';
 // import { competitions } from './pages/competitions';
 import { notebookPlugin } from './pages/notebook';
 // import { helpPlugin } from './pages/help';
-import { generateDefaultNotebookName } from './notebook-name';
+import { generateDefaultNotebookName, isNotebookEmpty } from './notebook-utils';
 import {
   IViewOnlyNotebookTracker,
   viewOnlyNotebookFactoryPlugin,
